Put Simplifier and MonadSimplify in one module
