+++ kore/src/Kore/Simplify/Data.hs
@@ -24,13 +24,7 @@ module Kore.Simplify.Data (
     MonadProf,
 ) where
 
-import Control.Monad.Catch (
-    MonadCatch,
-    MonadMask,
-    MonadThrow,
- )
 import Control.Monad.Reader
-import Control.Monad.State.Strict
 import Data.Map.Strict (
     Map,
  )
@@ -90,32 +84,6 @@ import SMT (
     SMT,
  )
 
--- * Simplifier
-
-data Env = Env
-    { metadataTools :: !(SmtMetadataTools Attribute.Symbol)
-    , simplifierCondition :: !(ConditionSimplifier Simplifier)
-    , simplifierAxioms :: !BuiltinAndAxiomSimplifierMap
-    , memo :: !(Memo.Self Simplifier)
-    , injSimplifier :: !InjSimplifier
-    , overloadSimplifier :: !OverloadSimplifier
-    , simplifierXSwitch :: !SimplifierXSwitch
-    }
-
-{- | @Simplifier@ represents a simplification action.
-
-A @Simplifier@ can send constraints to the SMT solver through 'MonadSMT'.
-
-A @Simplifier@ can write to the log through 'HasLog'.
--}
-newtype Simplifier a
-    = Simplifier (StateT SimplifierCache (ReaderT Env SMT) a)
-    deriving newtype (Functor, Applicative, Monad)
-    deriving newtype (MonadSMT, MonadLog, MonadProf)
-    deriving newtype (MonadIO, MonadCatch, MonadThrow, MonadMask)
-    deriving newtype (MonadReader Env)
-    deriving newtype (MonadState SimplifierCache)
-
 traceProfSimplify ::
     MonadProf prof =>
     Pattern RewritingVariableName ->
@@ -131,48 +99,6 @@ traceProfSimplify (Pattern.toTermLike -> termLike) =
             . Pretty.pretty
             <$> matchAxiomIdentifier termLike
 
-instance MonadSimplify Simplifier where
-    askMetadataTools = asks metadataTools
-    {-# INLINE askMetadataTools #-}
-
-    simplifyPattern sideCondition patt =
-        traceProfSimplify patt (Pattern.makeEvaluate sideCondition patt)
-    {-# INLINE simplifyPattern #-}
-
-    simplifyTerm = TermLike.simplify
-    {-# INLINE simplifyTerm #-}
-
-    simplifyCondition topCondition conditional = do
-        ConditionSimplifier simplify <- asks simplifierCondition
-        simplify topCondition conditional
-    {-# INLINE simplifyCondition #-}
-
-    askSimplifierAxioms = asks simplifierAxioms
-    {-# INLINE askSimplifierAxioms #-}
-
-    localSimplifierAxioms locally =
-        local $ \env@Env{simplifierAxioms} ->
-            env{simplifierAxioms = locally simplifierAxioms}
-    {-# INLINE localSimplifierAxioms #-}
-
-    askMemo = asks memo
-    {-# INLINE askMemo #-}
-
-    askInjSimplifier = asks injSimplifier
-    {-# INLINE askInjSimplifier #-}
-
-    askOverloadSimplifier = asks overloadSimplifier
-    {-# INLINE askOverloadSimplifier #-}
-
-    getCache = get
-    {-# INLINE getCache #-}
-
-    putCache = put
-    {-# INLINE putCache #-}
-
-    askSimplifierXSwitch = asks simplifierXSwitch
-    {-# INLINE askSimplifierXSwitch #-}
-
 -- | Run a simplification, returning the results along all branches.
 runSimplifierBranch ::
     Env ->
@@ -180,16 +106,6 @@ runSimplifierBranch ::
     LogicT Simplifier a ->
     SMT [a]
 runSimplifierBranch env = runSimplifier env . observeAllT
-
-{- | Run a simplification, returning the result of only one branch.
-
-__Warning__: @runSimplifier@ calls 'error' if the 'Simplifier' does not contain
-exactly one branch. Use 'evalSimplifierBranch' to evaluation simplifications
-that may branch.
--}
-runSimplifier :: Env -> Simplifier a -> SMT a
-runSimplifier env (Simplifier simplifier) =
-    runReaderT (evalStateT simplifier initCache) env
 
 mkSimplifierEnv ::
     SimplifierXSwitch ->
@@ -207,6 +123,8 @@ mkSimplifierEnv simplifierXSwitch verifiedModule sortGraph overloadGraph metadat
         Env
             { metadataTools = metadataTools
             , simplifierCondition
+            , simplifierTerm
+            , simplifierPattern
             , simplifierAxioms = earlySimplifierAxioms
             , memo = Memo.forgetful
             , injSimplifier
@@ -222,6 +140,12 @@ mkSimplifierEnv simplifierXSwitch verifiedModule sortGraph overloadGraph metadat
     simplifierCondition =
         {-# SCC "evalSimplifier/simplifierCondition" #-}
         Condition.create substitutionSimplifier
+    simplifierTerm =
+        {-# SCC "evalSimplifier/simplifierTerm" #-}
+        TermLike.simplify
+    simplifierPattern sideCondition patt =
+        {-# SCC "evalSimplifier/simplifierPattern" #-}
+        traceProfSimplify patt (Pattern.makeEvaluate sideCondition patt)
     -- Initialize without any builtin or axiom simplifiers.
     earlySimplifierAxioms = Map.empty
 
@@ -259,6 +183,8 @@ mkSimplifierEnv simplifierXSwitch verifiedModule sortGraph overloadGraph metadat
             Env
                 { metadataTools
                 , simplifierCondition
+                , simplifierTerm
+                , simplifierPattern
                 , simplifierAxioms
                 , memo
                 , injSimplifier
