+++ kore/test/Test/Kore/Rewrite/Axiom/EvaluationStrategy.hs
@@ -30,7 +30,7 @@ import Kore.Rewrite.Axiom.EvaluationStrategy
 import Kore.Rewrite.RewritingVariable (
     RewritingVariableName,
  )
-import Kore.Simplify.Simplify
+import Kore.Simplify.Simplify hiding (runSimplifier)
 import Kore.Simplify.Simplify qualified as AttemptedAxiom (
     AttemptedAxiom (..),
  )
