+++ kore/test/Test/Kore/Simplify/Application.hs
@@ -34,7 +34,7 @@ import Kore.Rewrite.RewritingVariable (
     mkConfigVariable,
  )
 import Kore.Simplify.Application
-import Kore.Simplify.Simplify
+import Kore.Simplify.Simplify hiding (runSimplifier)
 import Kore.Simplify.Simplify qualified as AttemptedAxiom (
     AttemptedAxiom (..),
  )
