+++ kore/test/Test/Kore/Simplify/Ceil.hs
@@ -37,7 +37,7 @@ import Kore.Simplify.Ceil qualified as Ceil (
     makeEvaluate,
     simplify,
  )
-import Kore.Simplify.Simplify
+import Kore.Simplify.Simplify hiding (runSimplifier)
 import Kore.Simplify.Simplify qualified as AttemptedAxiom (
     AttemptedAxiom (..),
  )
